Added functionality and created loadingScreen.css

+++ assets/js/script.js
@@ -1,27 +1,28 @@
 const delay = ms => new Promise(res => setTimeout(res, ms));
 
-const clear = document.querySelector("clearButton")
+const clear = document.querySelector("#clear")
 const hiddenDiv = document.getElementById("loadingScreen")
 
 
 //  Waits for two seconds before fading out of the running cat gif
 const loadContent = async () => {
-    await delay(2000);
     hiddenDiv.style.animation = "fade-out 2s forwards";
-    hiddenDiv.style.display = "hidden";
+    await delay(2000);
+    hiddenDiv.remove();
 }
 
 // Checks to see if the page is loaded, then runs the loadContent function
 document.addEventListener("DOMContentLoaded", loadContent())
 
 
-//  On click of a certain button on the form, will clear local storage. needs form creation though
+
+//  On click of a certain button on the form, will clear local storage.
 function clearForm() {
-    clear.addEventListener("onclick", function() {
+        if (window.confirm("Are you sure you want to start over?")) {
         localStorage.clear();
-        document.createTextNode("") // Fix to create element "Form cleared!"
-    })
-}
+        window.location.href = "/";
+        }
+    }
 
 function taskReminder() {
 // Renee
